utils/schema_to_xls/requirements.txt
psycopg2-binary==2.9.9
openpyxl==3.1.2
python-dotenv==1.0.1